chore: rewrite addClinic func and reimplement clinic service
