+++ src/main/java/com/example/ecommercebe/helper/mapper/MapperConfig.java
@@ -1,0 +1,12 @@
+package com.example.ecommercebe.helper.mapper;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MapperConfig {
+    @Bean
+    public ClinicMapper clinicMapper() {
+        return ClinicMapper.INSTANCE;
+    }
+}
