+++ src/main/resources/application.properties
@@ -2,9 +2,9 @@ spring.application.name=ecommerce-BE
 
 bezkoder.openapi.dev-url=http://localhost:8080
 
-spring.datasource.url=jdbc:mysql://localhost:3306/ecommerce_tasc
+spring.datasource.url=jdbc:mysql://localhost:3306/ecommerceproject
 spring.datasource.username=root
-spring.datasource.password=
+spring.datasource.password=Nam09082002@
 spring.jpa.show-sql=true
 
 #
